Changelog updates, moved files around.
