+++ CHANGELOG.md
@@ -1,62 +1,22 @@
-# GuitarModes
-Blending my passion for guitar with my desire to learn software development.
-
-## Overview
-I have been playing guitar for 13 years. Until this spring, I did not know any music theory beyond a few things:
-* Major / Minor Triads
-* Pentatonic Scale (Four-Five frets worth)
-  * Flattened Fifth Note
-* Major Scale from the Ionian Mode through the first octave of Phrygian.
-
-Recently, I have been learning music theory. It's dramatically changed the way I see the guitar. A ton of things are falling into place. (Learning songs, writing music, being able to think of my riffs in terms of notes, etc...)
-
-This project is the synthesis of my desires to learn software development and music theory. This application will provide the user with tools to practice scales and other music theory bits.
-
-## Plans
-### My plans from basic to incredibly ambitious are:
- - [x] Establish a way to generate a fretboard in plaintext that reflects a theoretical fretboard.
- - [x] Integrate hz and octaves, differentiate based on that instead of arbitrary index values.
- - [X] Modify getFretBoard.py such that it displays octaves of the notes (targeting version 0.23)
- - [ ] Modify getFretBoard such that it is object oriented.
- - [ ] Modify getFredboard.py such that the notes are all padded w/ spacing.
- - [ ] Establish clear definitions and use cases separating 'Aristotelian' Notes from 'Octave Defined' notes. (My language limiting my world, again.)
- - [ ] Dynamically emphasize different notes corresponding to different music theory concepts.
- - [ ] Implement ASIO integration--monitor input--come up with a very basic Rocksmith type thing for practicing.
- - [ ] Make a mobile version.
-
-### Examples:
-* What's a full tone higher than a C note?
-* To which major or minor scale does C Locrian belong?
-* To which scale does the following sequence of notes belong? 
- * C-D-E-F-G-A-B-C
-* The highlighted notes belong to which scale?
-* The highlighted notes belong to which mode?
-* What are the notes of the C Dorian Mode?
-
-## Anticipated Changes
-* Ability to use either sharp or flat notes. 
-  * Typically, Major keys' B, E, A, D, and G's accidentals are referred to by flats and not sharps.
-  * Typically, Minor keys' B and E's accidentals are referred to by flats and not sharps.
-* Depending on how I approach the algorithms to quiz you on key signatures / scale positions / modes...
-  * It's going to be a problem that the current system cannot account for octaves of notes. I may simply have to use frets as a fulcrum.
-
-## Changelog
-### Version 0.25 - 2022-12-
+# Changelog
+## Version 0.25 - 2022-12-
 * Updated Changelog
   * Changed dates from YYYY/MM/DD to YYYY-MM-DD.
   * Added some missing dates.
+  * Moved it to its own file: CHANGELOG.md
+* Moved py files into a lib directory
 
-### Version 0.24 - 2022-12-21
+## Version 0.24 - 2022-12-21
 * Improved Readability on code.
 * I neglected to include a version number, but it's PR #2
 
-### Version 0.23 - 2022-05-30
+## Version 0.23 - 2022-05-30
 * Fretboard now generates fretboard w/ Pitch Notation.
 * Updated goals to incldue making get_fretboard object oriented
   * This would allow for easily selecting different views.
   * It would probably be advisable to add a non Pitch Notation value/attribute to each note arrayIndex/object.
 
-### Version 0.22875 - 2022-05-29
+## Version 0.22875 - 2022-05-29
 * Achilles Paradox
 * Renamed get_note_library.py to pitch_notation
   * Refactored. Reduced redundancy. Made it easier to read.
@@ -65,16 +25,16 @@ This project is the synthesis of my desires to learn software development and mu
     * Helmholtz Pitch Notation
     * Scientific Pitch Notation
 
-### Version 0.2275 - 2022-05-29
+## Version 0.2275 - 2022-05-29
 * Hilarious version numbers
 * Cleaned up resources. (Modes, Keys, Alphabet)
 * Help from M.S. using delegation to clean up the methods in generateNoteLibrary.py
 
-### Version 0.225 - 2022-05-10
+## Version 0.225 - 2022-05-10
 * Fixed the getScale.py function to not produce unintelligble messes of output
 * Added pretty print and better naming of output.
 * Updated goals.
-#### Sample Output from getScale.py
+### Sample Output from getScale.py
 ```
  G minor Key Signature: ['G', 'A', 'A♯/B♭', 'C', 'D', 'D♯/E♭', 'F', 'G']
     G Ionian
@@ -93,13 +53,13 @@ This project is the synthesis of my desires to learn software development and mu
        ['F', 'G', 'A', 'A♯/B♭', 'C', 'D', 'D♯/E♭', 'F']
 ```
 
-### Verson 0.22 - 2022-05-10
+## Verson 0.22 - 2022-05-10
 * Implemented a means to generate hemholtz name
 * Implemented method that makes the generation of the hemholtz and scientific names conform to DRY principles
   * Factored out the common factor between hemholtz and scientific so its abstract
 
-#### Sample code from the fretboard generation
-##### Note that version 0.23 will have octave numerals included in the names:
+### Sample code from the fretboard generation
+#### Note that version 0.23 will have octave numerals included in the names:
 ```
 ['C', 'C♯/D♭', 'D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C', 'C♯/D♭', 'D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C']
 ['G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C', 'C♯/D♭', 'D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C', 'C♯/D♭', 'D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G']
@@ -109,12 +69,12 @@ This project is the synthesis of my desires to learn software development and mu
 ['D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C', 'C♯/D♭', 'D', 'D♯/E♭', 'E', 'F', 'F♯/G♭', 'G', 'G♯/A♭', 'A', 'A♯/B♭', 'B', 'C', 'C♯/D♭', 'D']
 ```
 
-### Version 0.21 2022-05-10
+## Version 0.21 2022-05-10
 * Renamed versioning scheme.
   * Added a digit following the decimal.
 * First version 0.2X
 * Implemented a means to generate the code to store the names of everything in a list.
-  #### Sample Output
+  ### Sample Output
     ```
     ...
     [36, "", "G#3/Ab3", 207.65234878997256],
@@ -133,14 +93,14 @@ This project is the synthesis of my desires to learn software development and mu
     ...
     ```
 
-### Version 0.15 - 2022-05-08
+## Version 0.15 - 2022-05-08
 * First Working Version
   * I'm going to break this going forward.
   * I need to implement a method to differentiate between keys via hz and octave. Currenly everything is just relative to some indexes I came up with.
   * More information here: https://www.ece.iastate.edu/~alexs/classes/2016_Spring_575/HW/HW5/files/piano-key-freq-wikipedia.pdf 
 
 
-### Version 0.14 - 2022-04-19
+## Version 0.14 - 2022-04-19
 * Neglected changelog for three weeks
 * Refactored some bits.
 * Added the ability to generate modes and select them.
@@ -166,7 +126,7 @@ E Locrian
 ['E', 'F', 'G', 'A', 'A#', 'C', 'D', 'E']
 ```
 
-### Version 0.13 - 2022-04-01
+## Version 0.13 - 2022-04-01
 * Implemented intonation flag.
 * I'm tired
 
@@ -179,7 +139,7 @@ Fb minor
 ```
 
 
-### Version 0.12 - 2022-04-01
+## Version 0.12 - 2022-04-01
 * renamed python files for fretboard and scales
 * created a functional generator for scales of both major and minor, without a working flag for sharpness
 ```
@@ -216,7 +176,7 @@ MINOR SCALES
          ['G', 'A', 'A#', 'C', 'D', 'D#', 'F', 'G']
 ```
 
-### Version 0.11 - 2022-03-31
+## Version 0.11 - 2022-03-31
 Everything you see here:
 * Goals described
 * Made a method to generate all the notes of a fretboard
