+++ lib/alphabet.py
