+++ lib/modes.py
