+++ lib/scales.py
